Int to Long in GQL queries (JS)

+++ sdks/js/packages/spark-sdk/src/graphql/mutations/RequestSwapLeaves.ts
@@ -3,9 +3,9 @@ import { FRAGMENT as LeavesSwapRequestFragment } from "../objects/LeavesSwapRequ
 export const RequestSwapLeaves = `
   mutation RequestSwapLeaves(
     $adaptor_pubkey: String!
-    $total_amount_sats: Int!
-    $target_amount_sats: Int!
-    $fee_sats: Int!
+    $total_amount_sats: Long!
+    $target_amount_sats: Long!
+    $fee_sats: Long!
     $user_leaves: [UserLeafInput!]!
   ) {
     request_leaves_swap(input: {
